move execution to a submodule
